cosmosdb-time-trigger/function/requirements.txt
azure-functions
feedparse
